pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "popkit"
version = "1.0.0"
description = "PopKit - LLM-agnostic developer orchestration engine"
authors = ["Joseph Cannon <joseph@thehouseofdeals.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "popkit"}]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[tool.poetry.urls]
"Homepage" = "https://github.com/jrc1883/popkit-claude"
"Repository" = "https://github.com/jrc1883/popkit-claude"

[tool.poetry.dependencies]
python = "^3.11"
popkit-shared = "^1.0.0"
popkit-cli = "^1.0.0"
popkit-mcp = {version = "^1.0.0", optional = true}
pyyaml = "^6.0"
filelock = "^3.13.0"
lxml = {version = "^6.0", optional = true}

[tool.poetry.extras]
# pip install popkit[mcp]  — includes MCP server
mcp = ["popkit-mcp"]
# pip install popkit[full]  — everything
full = ["popkit-mcp", "lxml"]

[tool.poetry.scripts]
popkit = "popkit_cli.main:main"
